Modified: Move Testcontainers logic to `MongoExtension`
